othello trainer, some minor consistency changes, rename resnet to LZ

+++ envs/_2048/vectmcts.py
@@ -1,12 +1,12 @@
 
 import torch
-from core.vz_resnet import VZResnet
+from core.lz_resnet import LZResnet
 
 from core.lazy_mcts import VectorizedLazyMCTS
 from envs._2048.vectenv import _2048Env
 
 class _2048LazyMCTS(VectorizedLazyMCTS):
-    def __init__(self, env: _2048Env, puct_c: int) -> None:
+    def __init__(self, env: _2048Env, puct_c: float) -> None:
         super().__init__(env, puct_c, 1e5)
